discordBot/requirements.txt
discord.py
motor
sentry-sdk
aiohttp
dnspython